Fixes for FZF and the setup script
FZFs bin directory wasn't included on the path.  Typically this is
a linux only issue as FZF is also installed using brew on the Mac.

The setup script had a broken check for lsb_release which stop a
call to apt to install some basic bits.  Also the version of NeoVim
in Ubuntus own repos is out of date so I had to add the repo for
NeoVims PPA to get an updated version of NeoVim.

Currently AstroNVim doesn't work entirely correctly under ARM due
to the way the NeoVim Ubuntu binary is built as it lacks support for
LuaJIT.